Add a `Gemfile` to manage ruby build dependencies
This commit is the same as toggl/webapp#ad204a30de17414e87d8a09acc81c63436bf4b91.
The message is duplicated here for reference.

This adds a `Gemfile` to control the ruby dependencies from the grunt
tasks. Namely, this ensures the developers are using `compass` at
version 0.12.* rather than the latest version 1.*.

This is required by the `grunt-contrib-compass` task in the required
version (0.7.1) [1], even though it isn't by the latest version (1.0.1
at this time) [2].

Developers can now deterministically get a working build environment
with:

    bundle
    npm install
    grunt

[1] - https://github.com/gruntjs/grunt-contrib-compass/tree/506a4dc09f7be09d9aa05eaf77aa1930eedcb67b#compass-task
[2] - https://github.com/gruntjs/grunt-contrib-compass#compass-task