Set-up watch and livereload for grunt
This (coupled with a small modification to the connect server task),
will enable us to do live editing of the website, seeing the changes on
the browser as we make them. Ideally, when a file changes, only the
least amount of work necessary would be done. This isn't currently the
case, but could be added later on.