Set-up a proxy server for `next.toggl.com/api`
This mimicks the set-up in toggl/webapp so that requests made to the
connect server serving the website that match the `/api/v.` format get
proxied to `next.toggl.com`. I would have skipped this by making the
requests directly from the browser to this staging API, but I had a
bunch of CORS problems, which wouldn't be easily solved without me
knowing how the server is set-up.

It also serves `dist/index.html` on a 404 error (this could be
improved by filtering out non `GET` requests). This fixes the
development environment's behaviour when a "virtual" route is reloaded
(i.e. localhost:9001/signup).