Make the API wrapper take an optional `endpoint`
This is just so we can support hitting both `https://toggl.com/api/v8`
or `https://next.toggl.com/api/v8`. I personally don't like constructor
params at all and think an options object is a much cleaner API for
this.

The commit also fixes the `TogglApi.prototype.request` method to
actually use the `method` parameter in its AJAX request (though the API
for this method could also be improved a bit, to make it more than just
a level of indirection from the `$.ajax` function) and the
`page/signup.coffee` view to make its request against the current host -
which is our grunt proxy to `next.toggl.com` - and send `data.email`
instead of `data.username`.

It'd be nice to start writing tests for everything, which I should do
sometime soon. This makes the core sign-up logic work.