Set-up Karma and add basic tests for `form-data`
This is mostly mimicking `toggl/webapp` with a few updates and
additions. I hope to be able to do both unit and E2E tests. I'm also
just spending time in this because it's already 11pm and it can't really
hurt to have tests. I was playing around with DevTools and looking for
information on load times and such, and I feel like tests make tweaking
the code for things like that much faster.