Start drafting tests for `api`
This adds a non-working/incomplete test suite to the `api` module. The
idea is to use the `pretender` module to mock the API responses and run
the tests in an isolated browser environment (though I don't think I'll
be able to get `pretender` to work on `browserify` without patches).

Once the structure is in place, it could be reused to write a full
client-side mock server.