Add basic Google Sign-Up implementation
This adds basic support for signing-up using Google. It mostly copies
the flow in the previous implementation (studied from `master`) and
tries to get around things that are already configured (such as the
redirect_url). We could tweak this in the front-end and make the Router
handle the code, rather than parsing it from the QueryString on our
second step.