Add initial experimental gulp build system
It's kind of late and this commit won't be permanent, so I won't
elaborate too much on what this is. The idea is to refactor the build
system, using Gulp (or not necessarily - though that's what I did).

This build system has support for source maps (from all the input files,
CSS, Coffee-Script, `node_modules` into the `vendor.js` bundle etc.).

It also has support for LiveReload, proxying and most of the things that
we're previously supported in the Grunt configuration.

The main issue right now is that for some reason this is slow (2x slower
than Grunt when we're generating source-maps and minifying)... As in
really slow. I think the coffee-script compilation as a Browserify
transform makes the browserify task really slow and that's a little hard
to fix. We can move compilation a step up, but then I'm not sure how the
source maps will be generated.

I'm going to try to decouple the coffee-script builds from the
browserify bundling and add caching, which will probably speed things
up.