Fix closing the login modal when on `/login`
This issue was related to a really weird bug. When the page was first
loaded, `LoginPopup::render` was called once. This call (in
`View::render`) executed the `Modal::preRender` hook, where an unique
element is created for the popup and event binding happens.

The `View` class however triggers its `render` method every time a
`change` event is triggered on its Model. So when the
`/api/v8/me/logged` request responded, `render` was called again.

This caused the screen to flicker and `loginPopup.$el` to be set again.

Waiting for the `change` event fixes that, but I'm thinking that
`Modal::preRender` should check if `@$el` exists before creating another
element and overwriting itself. After all, the best UX would be to
render the login popup right away, but show the user he/she is logged-in
from the `LoginPopup` if/when the response comes in.