Use the bleeding edge version of Backbone
This closes toggl/webapp#1105.

Because of usage of `window.location.hash`, when at "root" routes, the
`Backbone.History.prototype.start` method reseted our Google redirect
routes on IE [1]. This is because on Chrome/Firefox/etc., if we have a
route that ends with a trailing hash (`/?something=else#`),
`window.location.hash` will be `''`, but on IE, it'll be `'#'`. There
was an existence test over `window.location.hash` which triggered the
`History` object to try to turn hash based URLs into `pushState` based
URLs [2]. This was fixed in master, so we'll be using the latest version
until `1.1.3` is published [3].

[1] - https://github.com/jashkenas/backbone/blob/53f77901a4ea9c7cf75d3db93ddddf491998d90f/backbone.js#L1438
[2] - https://github.com/jashkenas/backbone/blob/53f77901a4ea9c7cf75d3db93ddddf491998d90f/backbone.js#L1436-L1441
[3] - https://github.com/jashkenas/backbone/commit/d52b5f92ae57a8c3ebbce3983a65a80e9d75900c