more info for a sign up fail alert
Failure to sign up due to 'account already exists' + 'google sign in
is disabled' now throws a more helpful / user friendly alert message,
that even includes a link/url to the appropiate knowledge base article:
http://support.toggl.com/google-sign-in/.

In the future, instead of making the user copy/paste the url in case
they are interested, we could have a button that links to the article
directly, via using a custom alert. SweetAlert was proposed during a
Slack discussion, and it looks very good indeed:
http://t4t5.github.io/sweetalert/.

Changes in the error messages backend provides break this 'hack', error
code constants (maybe via a response header so that response text
remains plain) could be considered.