Change OBMu72 excluded behaviour
They are now redirected to the webapp straightaway, instead of going
through the Overview.

We are making this change since the test results for OBMu62 are flawed
(the first experiment about the Overview page) - both included and
excluded groups were saw the Overview page in that test. A/A test.