Fix nav not taking into account initial user state
If we assumed the user is logged in from the beginning, the UI wasn't
reflecting that. And wouldn't update it later, since there is no change to the
`logged` property.

Besides updating the UI when the userState's `logged` property changes,
it now correctly reflects the userState after initially rendering the UI.