Fix canonical tag
Before it always had an `href` value of `https://toggl.com`, so all
crawlers were told to redirect to `https://toggl.com` when viewing one
of the `website` app pages.

`website` app pages now only have a canonical tag it their url has 'www'
in it, and it points to the same url but without 'www.'

Related to #477.