Use toggl.space for dev API
Adds secure: false to the proxy configuration so it wouldn't check for
the certificates. This seems to be the way to go with
grunt-connect-proxy.

Also updates grunt-connect-proxy to a version that supports this flag
(this is a later version than the current most recent version).
Make sure you update grunt-connect-proxy for this to work!