Add command for building large assets
And have build rewrite the urls to assets.toggl.com and omit the actual
assets from the build.

`grunt build:largeAssets` builds the large assets under
`dist-assets/` because there is no need to ship them as part of the
dist build but we still want to minify them for the deployment script.

closes #682